1) Переписан класс transition; 2) Изменено название папки с классом state.
