+++ src/renderer/src/lib/drawable/TransitionNode/ArrowsWithoutLabel.ts
@@ -1,0 +1,47 @@
+import { CanvasEditor } from '@renderer/lib/CanvasEditor';
+import { Note, Transition } from '@renderer/lib/drawable';
+import { transitionStyle } from '@renderer/lib/styles';
+import { Drawable } from '@renderer/lib/types/drawable';
+import {
+  degrees_to_radians,
+  drawCircle,
+  drawCurvedLine,
+  drawTriangle,
+  getLine,
+} from '@renderer/lib/utils';
+
+/**
+ * Выполняет отрисовку стрелки между двумя(тремя) движущимися блоками:
+ */
+export class ArrowsWithoutLabel implements Drawable {
+  constructor(private parent: Transition, private app: CanvasEditor) {}
+
+  draw(ctx: CanvasRenderingContext2D) {
+    const targetBounds = this.parent.target.drawBounds;
+    const sourceBounds = this.parent.source.drawBounds;
+
+    const line = getLine(
+      { ...targetBounds, height: targetBounds.height + targetBounds.childrenHeight },
+      { ...sourceBounds, height: sourceBounds.height + sourceBounds.childrenHeight },
+      10
+    );
+
+    const data = this.parent.data;
+
+    ctx.lineWidth = transitionStyle.width;
+    ctx.strokeStyle = data.color;
+    ctx.fillStyle = data.color;
+
+    if (!this.parent.isSelected && this.parent.source instanceof Note) {
+      ctx.globalAlpha = 0.3;
+    }
+    drawCurvedLine(ctx, line, 12 / this.app.model.data.scale);
+    ctx.globalAlpha = 1;
+    drawCircle(ctx, {
+      position: line.end,
+      radius: transitionStyle.startSize / this.app.model.data.scale,
+      fillStyle: data.color,
+    });
+    drawTriangle(ctx, line.start, 10 / this.app.model.data.scale, degrees_to_radians(line.se));
+  }
+}
