+++ src/renderer/src/lib/drawable/index.ts
@@ -1,8 +1,8 @@
-export * from './Node';
+export * from './StateNode';
 export * from './Children';
 export * from './Events';
 export * from './GhostTransition';
 export * from './Note';
 export * from './Picto';
 export * from './Shape';
-export * from './Transition';
+export * from './TransitionNode';
